Add files via upload
My project focuses on implementing a Data Vault architecture for the Pubs database and its associated data marts. The goal is to enhance the data integration, historical tracking, and scalability of the Pubs database, ensuring that the data storage is both flexible and robust. By employing the Data Vault methodology, the project will systematically capture and store all historical data changes in a way that supports agile business intelligence and analytics. This approach not only preserves data integrity and auditability but also simplifies the integration of new data sources and business rules. The final solution will provide a comprehensive, reliable, and scalable data warehouse environment tailored for the dynamic needs of the publishing industry.